requirements.txt
termcolor
pynput
pyreadline
web3==5.31.1
py-solc-x
pyevmasm
